backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
sendgrid==6.11.0
supabase==2.3.0
schedule==1.2.0
